move: Name클래스 Member패키지로 이동
